Signin/requirements.txt
fastapi
uvicorn
pymssql
azure.identity
azure.keyvault
python-dotenv
twilio
pyotp
python-jose
redis